Final project V1.0

+++ 7thModule/Programação Web 3/final-project-api/src/product/product.service.ts
@@ -1,7 +1,11 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
-import { CreateProductDto } from './dto/create-product.dto';
-import { UpdateProductDto } from './dto/update-product.dto';
+import { CreateProductDto } from './dto/product.dto';
+import { UpdateProductDto } from './dto/product.dto';
 import { Product } from '@prisma/client';
 
 @Injectable()
@@ -9,6 +13,7 @@ export class ProductService {
   constructor(private prisma: PrismaService) {}
 
   async create(data: CreateProductDto): Promise<Product> {
+    this.validateProductData(data);
     return this.prisma.product.create({ data });
   }
 
@@ -17,14 +22,47 @@ export class ProductService {
   }
 
   async findOne(id: number): Promise<Product> {
-    return this.prisma.product.findUnique({ where: { id } });
+    const product = await this.prisma.product.findUnique({ where: { id } });
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
+    return product;
   }
 
   async update(id: number, data: UpdateProductDto): Promise<Product> {
+    const product = await this.prisma.product.findUnique({ where: { id } });
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
+    this.validateProductData(data);
     return this.prisma.product.update({ where: { id }, data });
   }
 
   async remove(id: number): Promise<Product> {
+    const product = await this.prisma.product.findUnique({ where: { id } });
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
     return this.prisma.product.delete({ where: { id } });
+  }
+
+  private validateProductData(data: CreateProductDto | UpdateProductDto) {
+    if (typeof data.price !== 'number' || isNaN(data.price) || data.price < 0) {
+      throw new BadRequestException('Invalid price value');
+    }
+    if (
+      typeof data.userId !== 'number' ||
+      isNaN(data.userId) ||
+      data.userId <= 0
+    ) {
+      throw new BadRequestException('Invalid userId value');
+    }
+    if (
+      typeof data.categoryId !== 'number' ||
+      isNaN(data.categoryId) ||
+      data.categoryId <= 0
+    ) {
+      throw new BadRequestException('Invalid categoryId value');
+    }
   }
 }
